+++ CHANGELOG.md
@@ -10,6 +10,11 @@ All notable changes to this branch will be documented in this file.
 - Resolved an issue where the app would crash upon attempting to close it while the `Home` page was actively updating system usage statistics.
 - Fixed an issue where some failed apps were being displayed as successfully uninstalled.
 
+### Changes
+
+- Updated the toggle switch state management in `FeaturesPage`, `OptimizeSystemPage`, and `PrivacyPage` to use Windows Registry entries instead of `LocalSettings` for saving and retrieving states.
+
+
 ## 1.3.0 - Released
 
 ### Added